+++ src/GUI/right_panel.py
@@ -1,5 +1,5 @@
 from PySide6.QtWidgets import (
-    QWidget, QLabel, QTextEdit, QPushButton, QGridLayout, QSizePolicy, QStackedWidget, QPlainTextEdit, QVBoxLayout, QLineEdit, QFrame, QHBoxLayout, QMessageBox
+    QWidget, QLabel, QTextEdit, QPushButton, QGridLayout, QSizePolicy, QStackedWidget, QPlainTextEdit, QVBoxLayout, QLineEdit, QFrame, QHBoxLayout, QMessageBox, QComboBox
 )
 from PySide6.QtCore import Qt
 import threading 
@@ -19,17 +19,11 @@ class RightPanel(QWidget):
         # Crearea layout-ului principal
         self.main_layout = QVBoxLayout(self)
         self.setLayout(self.main_layout)
-        
-        # Crearea butoanelor de taburi
-        self.tab_buttons = QWidget()
-        self.tab_buttons_layout = QGridLayout(self.tab_buttons)
-        self.main_layout.addWidget(self.tab_buttons)
-        
-        self.testing_button = QPushButton("Testing", self, clicked=self.show_testing_tab)
-        self.server_button = QPushButton("Server", self, clicked=self.show_server_tab)
-        
-        self.tab_buttons_layout.addWidget(self.testing_button, 0, 0)
-        self.tab_buttons_layout.addWidget(self.server_button, 0, 1)
+
+        self.functions = QComboBox(self)
+        self.functions.addItems(["Testing", "Server"])
+        self.functions.currentIndexChanged.connect(self.toggle_tabs)
+        self.main_layout.addWidget(self.functions)
         
         # Crearea QStackedWidget pentru a comuta între tab-uri
         self.stacked_widget = QStackedWidget()
@@ -151,6 +145,13 @@ class RightPanel(QWidget):
             self.config = json.load(file)
         self.user_name = self.config.get('profile_name')
 
+    def toggle_tabs(self):
+        curr_funct = self.functions.currentText()
+        if curr_funct == "Testing":
+            self.show_testing_tab()
+        elif curr_funct == "Server":
+            self.show_server_tab()
+
     def start_server_option(self):
         if self.password_entry.text() == "":
             QMessageBox.information(self, "Info", "You need to enter the password!")
@@ -226,7 +227,13 @@ class RightPanel(QWidget):
             background-color: {theme.get("background_color")};
             color: {theme.get("text_color")};
         """)
-
+        self.functions.setStyleSheet(f"""
+                                     color: {theme.get('text_color')};
+                                     background-color: {theme.get('editor_background')};
+                                     border: 1px solid {theme.get('border_color')};
+                                     font-size: 14px;
+                                     padding: 4px;
+                                     """)
         self.input_label.setStyleSheet(f"color: {theme.get('text_color')};")
         self.output_label.setStyleSheet(f"color: {theme.get('text_color')};")
         self.expected_label.setStyleSheet(f"color: {theme.get('text_color')};")
@@ -278,8 +285,6 @@ class RightPanel(QWidget):
         self.diff.setStyleSheet(button_style)
         self.fetch.setStyleSheet(button_style)
         self.send_button.setStyleSheet(button_style)
-        self.testing_button.setStyleSheet(button_style)
-        self.server_button.setStyleSheet(button_style)
         self.connect_button.setStyleSheet(button_style)
         self.disconnect_button.setStyleSheet(button_style)
         self.start_server.setStyleSheet(button_style)