+++ src/GUI/kilo_tools.py
@@ -1,4 +1,4 @@
-from PySide6.QtWidgets import QApplication, QMainWindow, QLabel, QWidget
+from PySide6.QtWidgets import QLabel, QWidget
 from PySide6.QtCore import Qt
 from PySide6.QtGui import QIcon, QFont