+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Version 2.1
 
+### Major improvements:
+
+
 ### Added:
 
 1. Inbox
 2. FOR-{} completions e.g FOR-J, FOR-X -> int j, int x -> for()
 3. Author Details -> Templates
+4. New tab switch mode in right frame
 
 ### Fixed/Changes: