+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### Major improvements:
 
-1. Problem viewer
+1. Problem viewer -> mini google inside the app
 
 ### Added:
 
@@ -13,11 +13,13 @@
 5. "neutral-green.json" theme
 6. "vulcano.json" theme
 7. Credits in config.json 
+8. Tool bar in left frame
 
 ### Fixed/Changes:
 
 1. Highlighter improvement 
 2. Changed the pbinfo interface
+3. Fixed resized window bug
 
 # Version 2.0